+++ src/blocks.js
@@ -1723,6 +1723,7 @@ SyntaxElementMorph.prototype.labelPart = function (spec) {
 
         // Video motion
 
+/* // +++ to be removed
         case '%vid': // video modes
             part = new InputSlotMorph(
                 null,
@@ -1734,7 +1735,8 @@ SyntaxElementMorph.prototype.labelPart = function (spec) {
                 true
             );
             break;
-        case '%motype':
+*/
+        case '%vid':
             part = new InputSlotMorph(
                 null,
                 false, {
@@ -2407,7 +2409,7 @@ SyntaxElementMorph.prototype.endLayout = function () {
     %f      - round function slot, unevaluated if replaced,
     %r      - round reporter slot
     %p      - hexagonal predicate slot
-    %vid    - chameleon colored rectangular drop-down for video modes
+    %vid    - chameleon colored rectangular drop-down for video modes // +++ check these
     %motype - chameleon colored rectangular drop-down for motion type
     %on     - chameleon colored rectangular drop-down for motion detection scope