+++ HISTORY.md
@@ -80,6 +80,7 @@
 ### 2019-05-08
 * Blocks, Objects, Threads: integrated video capture control into global settings prims in Sensing
 * Blocks, Threads: added a %self menu
+* Blocks, Threads, Objects: Finalized Video ops
 
 ### 2019-05-07
 * Blocks, Objects, Threads, Video: optimized video motion detection